Added Database Files
Added several database scripts. One is drop table and create, basicall only run this to update the schema. If there is any data in the tables, it will be lost.

The next is to create the tables if they don't already exist.

The last is all of the stored procedures.